Main Code

+++ WeatherAPI/WeatherForecast.cs
@@ -2,7 +2,7 @@
 {
     public class WeatherForecast : iWeather
     {
-        public int temp { get; set; }
+        public double temp { get; set; }
         public int winddegrees { get; set; }
         public string weather { get; set; }
