+++ WeatherAPI/iWeather.cs
@@ -2,7 +2,7 @@
 {
     interface iWeather
     {
-        static int temp;
+        static double temp;
         static int winddegrees;
         static string winddirection;
         static string weather;
